Added changes of entity1

+++ employee-management-hrms/employee-management-hrms/src/main/java/com/employee/management/hrms/repository/LeaveRequestRepository.java
@@ -7,13 +7,13 @@ import org.springframework.stereotype.Repository;
 import com.employee.management.hrms.entity.LeaveRequest;
 import com.employee.management.hrms.entity.LeaveStatus;
 @Repository
-public interface LeaveRequestRepository extends JpaRepository<LeaveRequest, Long> {
+public interface LeaveRequestRepository extends JpaRepository<LeaveRequest, Integer> {
 
-    List<LeaveRequest> findByEmployee_EmployeeId(Long employeeId);
+    List<LeaveRequest> findByEmployee_EmployeeId(Integer employeeId);
 
     List<LeaveRequest> findByStatus(LeaveStatus status);
 
-    List<LeaveRequest> findByApprovedBy_EmployeeId(Long managerId);
+    List<LeaveRequest> findByApprovedBy_EmployeeId(Integer managerId);
 
-    List<LeaveRequest> findByEmployee_Department_DepartmentId(Long departmentId);
+    List<LeaveRequest> findByEmployee_Department_DepartmentId(Integer departmentId);
 }
